WiFi provisioning started and working! still some rough edges that needs to be cleaned!! -> wrong password: DONE but maybe still little primitive => THIS IS OK -> provisioning after button press (factory reset): NOT DONE
HTTP connection implemented. There is now a connection possible between the server and the esp32. Over this connection we are able to send data in different kind of request!
DONE: variable content for HTTP message
DONE: put data in JSON!
DONE: make program better for working in overall scenarios.
DONE: being able to add data to the database!!!

SPI configuration started but not finished!
LED on board is one when provisioning is in progress!