WiFi provisioning started and working! still some rough edges that needs to be cleaned!! -> wrong password: DONE but maybe still little primitive => THIS IS OK -> provisioning after button press (factory reset): DONE -> LED on board will blink when provisioning and stay ON when provisioning is successful.
HTTP connection implemented. There is now a connection possible between the server and the esp32. Over this connection we are able to send data in different kind of request!
DONE: variable content for HTTP message
DONE: put data in JSON!
DONE: make program better for working in overall scenarios.
DONE: being able to add data to the database!!!

Shared data structure:
DONE: implementation of simple data structure with a primitive reader and writer task
TODO: Optimization of the task

SPI configuration started but not finished!